# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_SamandarMadaliev_ex-rate_internal_http_schemas.CreateRateRequest:
    properties:
      base_currency:
        type: string
      quote_currency:
        type: string
    type: object
  github_com_SamandarMadaliev_ex-rate_internal_http_schemas.CreateRateResponse:
    properties:
      id:
        type: string
    type: object
  github_com_SamandarMadaliev_ex-rate_internal_http_schemas.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  github_com_SamandarMadaliev_ex-rate_internal_http_schemas.RateResponse:
    properties:
      base_currency:
        type: string
      created_at:
        type: string
      id:
        type: string
      price:
        type: number
      price_timestamp:
        type: string
      quote_currency:
        type: string
      status:
        $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_models.RateStatus'
      updated_at:
        type: string
    type: object
  github_com_SamandarMadaliev_ex-rate_internal_models.Currency:
    properties:
      id:
        type: integer
      slug:
        type: string
    type: object
  github_com_SamandarMadaliev_ex-rate_internal_models.RateStatus:
    enum:
    - pending
    - completed
    - failed
    type: string
    x-enum-varnames:
    - RateStatusPending
    - RateStatusCompleted
    - RateStatusFailed
info:
  contact: {}
  description: Currency exchange rate service that fetches and stores rates for currency
    pairs.
  title: Ex-Rate API
  version: "1.0"
paths:
  /currencies:
    get:
      description: Returns all currencies known to the service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_models.Currency'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.ErrorResponse'
      summary: List currencies
      tags:
      - currencies
  /health:
    get:
      description: Pings the database to verify the service is ready to serve traffic
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.ErrorResponse'
      summary: Health check
      tags:
      - health
  /rates:
    post:
      consumes:
      - application/json
      description: Creates a pending rate row for a currency pair and enqueues a background
        job to fetch its price
      parameters:
      - description: Currency pair
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.CreateRateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.CreateRateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.ErrorResponse'
      summary: Create a rate
      tags:
      - rates
  /rates/{id}:
    get:
      description: Returns the rate row for the given ID
      parameters:
      - description: Rate ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.RateResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.ErrorResponse'
      summary: Get a rate by ID
      tags:
      - rates
  /rates/latest:
    get:
      description: Returns the most recently created rate for the given base/quote
        currency IDs
      parameters:
      - description: Base currency ID
        in: query
        name: base
        required: true
        type: integer
      - description: Quote currency ID
        in: query
        name: quote
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.RateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_SamandarMadaliev_ex-rate_internal_http_schemas.ErrorResponse'
      summary: Get the latest rate for a pair
      tags:
      - rates
swagger: "2.0"